# Perf: add an software override (1.5sec) for `TimeoutPropose`